+++ @navikt/aksel/src/codemod/transforms/v8.0.0/box/tests/index.test.ts
@@ -10,6 +10,7 @@ const fixtures = [
   "box-radius",
   "box-simple",
   "expression-container",
+  "real-examples",
 ];
 
 for (const fixture of fixtures) {