+++ @navikt/aksel/src/codemod/transforms/v8.0.0/primitives-spacing/tests/spacing.test.ts
@@ -8,6 +8,7 @@ const fixtures = [
   "subComponent",
   "idempotent",
   "directImport",
+  "composed",
 ];
 
 for (const fixture of fixtures) {